DDF-1565 Changed Managed Service Factories from Spring to Blueprint. Affected Packages : content-core-directorymonitor catalog-opensearch-source spatial-csw-source spatial-kml-transformer spatial-wfs-mapper spatial-wfs-1_0_0-source spatial-wfs-2_0_0-source platform-mime-core-configureableresolver platform-scheduler

+++ catalog/spatial/csw/spatial-csw-source/src/test/java/org/codice/ddf/spatial/ogc/csw/catalog/source/TestCswSource.java
@@ -787,8 +787,8 @@ public class TestCswSource extends TestCswSourceBase {
         configuration.put("receiveTimeout", 10000);
         cswSource.refresh(configuration);
 
-        assertEquals(cswSource.getConnectionTimeout(), 10000);
-        assertEquals(cswSource.getReceiveTimeout(), 10000);
+        assertEquals(cswSource.getConnectionTimeout().intValue(), 10000);
+        assertEquals(cswSource.getReceiveTimeout().intValue(), 10000);
     }
 
     @Test
